// osa9/ts-react-app/src/types.tsx
interface CoursePartBase {
    name: string
    exerciseCount: number
}

interface DescribedCoursePart extends CoursePartBase {
    description: string
}

interface CoursePartOne extends DescribedCoursePart {
    name: 'Fundamentals'
}

interface CoursePartTwo extends CoursePartBase {
    name: 'Using props to pass data'
    groupProjectCount: number
}

interface CoursePartThree extends DescribedCoursePart {
    name: 'Deeper type usage'
    exerciseSubmissionLink: string
}

interface CoursePartFour extends DescribedCoursePart {
    name: 'Love for typescript'
}

export type CoursePart = CoursePartOne | CoursePartTwo | CoursePartThree | CoursePartFour